refactor(web): extract SessionExpiredCard and useConfigSave hook
- Extract duplicated 401 card from ChatForm / RunForm into
  components/SessionExpiredCard.tsx.
- Consolidate PUT /api/config + 422 validation + SaveStatus across
  Portfolio / Geopolitical / WatchSectors forms into lib/hooks/useConfigSave.ts.
- Net diff: -114 lines in screen components.
