+++ apps/web/components/screens/ChatForm.tsx
@@ -3,6 +3,7 @@ import { useCallback, useEffect, useRef, useState } from "react"
 import ReactMarkdown from "react-markdown"
 import remarkGfm from "remark-gfm"
 
+import { SessionExpiredCard } from "@/components/SessionExpiredCard"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
 import { LoadingDots } from "@/components/ui/loading-dots"
@@ -247,23 +248,7 @@ export function ChatForm() {
   }, [listening])
 
   if (sessionExpired) {
-    return (
-      <Card>
-        <CardContent
-          className="space-y-2 pt-6 text-sm"
-          data-testid="session-expired"
-        >
-          <p className="font-medium text-destructive">Session expired</p>
-          <p className="text-muted-foreground">
-            The bearer token in <code className="font-mono">apps/web/.token</code>
-            {" "}no longer matches{" "}
-            <code className="font-mono">~/.ai-agent/session-token</code>.
-            Restart the dev server (<code className="font-mono">bin/serve.sh</code>)
-            to mirror a fresh token and refresh this page.
-          </p>
-        </CardContent>
-      </Card>
-    )
+    return <SessionExpiredCard />
   }
 
   return (
