+++ apps/web/components/screens/RunForm.tsx
@@ -1,6 +1,7 @@
 "use client"
 import { useCallback, useEffect, useRef, useState } from "react"
 
+import { SessionExpiredCard } from "@/components/SessionExpiredCard"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
 import { LoadingDots } from "@/components/ui/loading-dots"
@@ -127,23 +128,7 @@ export function RunForm() {
   }, [dryRun])
 
   if (sessionExpired) {
-    return (
-      <Card>
-        <CardContent
-          className="space-y-2 pt-6 text-sm"
-          data-testid="session-expired"
-        >
-          <p className="font-medium text-destructive">Session expired</p>
-          <p className="text-muted-foreground">
-            The bearer token in <code className="font-mono">apps/web/.token</code>
-            {" "}no longer matches{" "}
-            <code className="font-mono">~/.ai-agent/session-token</code>.
-            Restart the dev server (<code className="font-mono">bin/serve.sh</code>)
-            to mirror a fresh token and refresh this page.
-          </p>
-        </CardContent>
-      </Card>
-    )
+    return <SessionExpiredCard />
   }
 
   const started = fmtTs(job?.started_at)
